pyproject.toml
[project]
name = "spectre"
version = "1.0.0"
description = "Weight Outlier Detection & Integrity Fingerprint Engine"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "torch>=2.0.0",
    "safetensors>=0.4.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "gudhi>=3.7.0",
    "pot>=0.9.0",
    "stumpy>=1.12.0",
    "networkx>=3.0.0",
    "pytest>=7.4.0",
]

[project.scripts]
spectre = "spectre.cli:main"
